### 2021-10-04 c7772e8
remove commented lines from bibliography
