+++ paper/ref.bib
@@ -28,7 +28,6 @@
 	pages     = {49--75},
 	publisher = {EasyChair},
 	year      = {2020},
-	@url       = {https://doi.org/10.29007/zkf6},
 	doi       = {10.29007/zkf6},
 	timestamp = {Sun, 15 Aug 2021 00:16:52 +0200},
 	biburl    = {https://dblp.org/rec/conf/arch/GerettiSABCCCFF20.bib},
@@ -57,7 +56,6 @@
 	pages     = {16--48},
 	publisher = {EasyChair},
 	year      = {2020},
-	@url       = {https://doi.org/10.29007/7dt2},
 	doi       = {10.29007/7dt2},
 	timestamp = {Sun, 15 Aug 2021 00:16:52 +0200},
 	biburl    = {https://dblp.org/rec/conf/arch/AlthoffBBFFFKLM20.bib},
@@ -78,7 +76,6 @@
 	publisher    = {Zenodo},
 	version      = {v0.6.16},
 	doi          = {10.5281/zenodo.4993670},
-	@url          = {https://doi.org/10.5281/zenodo.4993670}
 }
 
 @misc{legat2020mathoptinterface,
@@ -123,15 +120,12 @@
   author    = {Miriam Garc{\'{\i}}a Soto and
                Thomas A. Henzinger and
                Christian Schilling},
-  @editor    = {Sergiy Bogomolov and
-               Rapha{\"{e}}l M. Jungers},
   title     = {Synthesis of hybrid automata with affine dynamics from time-series
                data},
   booktitle = {{HSCC}},
   pages     = {2:1--2:11},
   publisher = {{ACM}},
   year      = {2021},
-  @url       = {https://doi.org/10.1145/3447928.3456704},
   doi       = {10.1145/3447928.3456704}
 }
 
@@ -170,7 +164,6 @@
   pages     = {1--6},
   publisher = {{IEEE}},
   year      = {2020},
-  @url       = {https://doi.org/10.1109/MEMOCODE51338.2020.9314994},
   doi       = {10.1109/MEMOCODE51338.2020.9314994}
 }
 
@@ -183,7 +176,6 @@
   pages={65--98},
   year={2017},
   publisher={SIAM},
-  @url       = {https://doi.org/10.1137/141000671},
   doi       = {10.1137/141000671}
 }
 
@@ -219,7 +211,6 @@
 	pages     = {41--50},
 	publisher = {{ACM}},
 	year      = {2018},
-	@url       = {https://doi.org/10.1145/3178126.3178128},
 	doi       = {10.1145/3178126.3178128}
 }
 
@@ -235,7 +226,6 @@
   volume    = {10227},
   pages     = {288--294},
   year      = {2017},
-  @url       = {https://doi.org/10.1007/978-3-319-57288-8_20},
   doi       = {10.1007/978-3-319-57288-8\_20}
 }
 
@@ -259,8 +249,6 @@
 
 @inproceedings{Althoff15,
   author    = {Matthias Althoff},
-  @editor    = {Goran Frehse and
-               Matthias Althoff},
   title     = {An Introduction to {CORA} 2015},
   booktitle = {ARCH@CPSWeek},
   series    = {EPiC Series in Computing},
@@ -268,7 +256,6 @@
   pages     = {120--151},
   publisher = {EasyChair},
   year      = {2015},
-  @url       = {https://doi.org/10.29007/zbkv},
   doi       = {10.29007/zbkv}
 }
 
@@ -283,8 +270,6 @@
                Antoine Girard and
                Thao Dang and
                Oded Maler},
-  @editor    = {Ganesh Gopalakrishnan and
-               Shaz Qadeer},
   title     = {{SpaceEx}: Scalable Verification of Hybrid Systems},
[…]
