+++ paper/header.tex
@@ -1,16 +1,16 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{LazySets.jl: Scalable Symbolic-Numeric Set Computations}
+\title{LazySets.jl: Scalable Symbolic-Numeric Set Computations$^*$}
 
-\author[1,*]{Marcelo Forets}
-\author[2,*]{Christian Schilling}
+\author[1]{Marcelo Forets}
+\author[2]{Christian Schilling}
 \affil[1]{Universidad de la Rep\'{u}blica, Uruguay}
 \affil[2]{University of Konstanz, Germany}
 
 \keywords{Set propagation, Geometry, Reachability analysis, Hybrid system, Support function, Formal verification}
 
 \hypersetup{
-pdftitle = {LazySets.jl: Scalable Symbolic-Numeric Set Computations},
+pdftitle = {LazySets.jl: Scalable Symbolic-Numeric Set Computations$^*$},
 pdfsubject = {JuliaCon 2019 Proceedings},
 pdfauthor = {Marcelo Forets, Christian Schilling},
 pdfkeywords = {Set propagation, Geometry, Reachability analysis, Hybrid system, Support function, Formal verification},
