### 2021-10-27 #97 @whedon
[REVIEW]: LazySets.jl: Scalable Symbolic-Numeric Set Computations
**Submitting author:** @mforets (<a href="http://orcid.org/0000-0002-9831-7801">Marcelo Forets</a>)
**Repository:** <a href="https://github.com/JuliaReach/LazySets-JuliaCon21" target ="_blank">https://github.com/JuliaReach/LazySets-JuliaCon21</a>
**Version:** v1.53.4
