+++ paper/ref.bib
@@ -129,14 +129,15 @@
   doi       = {10.1145/3447928.3456704}
 }
 
-@article{forets2021reachability,
-	title={Reachability of weakly nonlinear systems using {C}arleman linearization},
-	author={Forets, Marcelo and Schilling, Christian},
-	journal={CoRR},
-	year={2021},
-	url       = {https://arxiv.org/abs/2108.10390},
-	eprinttype = {arXiv},
-	eprint    = {2108.10390}
+@inproceedings{forets2021reachability,
+  author    = {Forets, Marcelo and Schilling, Christian},
+  title     = {Reachability of weakly nonlinear systems using {C}arleman linearization},
+  booktitle = {{RP}},
+  series    = {LNCS},
+  volume    = {13035},
+  pages     = {85--99},
+  year      = {2021},
+  doi       = {10.1007/978-3-030-89716-1\_6}
 }
 
 @article{schilling2021verification,
