+++ paper/applications.tex
@@ -297,7 +297,7 @@ Below we show how to use \href{https://github.com/JuliaPy/pyjulia}{pyjulia} for 
 %
 We can see that it is not necessary to use Julia objects everywhere; NumPy arrays can also be used, making the interoperability between Julia and Python effortless.
 
-\begin{minipage}{\linewidth}
+% \begin{minipage}{\linewidth}
 	\begin{lstlisting}[language=python]
 $ python3 -m pip install --user julia
 
@@ -319,4 +319,4 @@ $ python3
 >>> volume(B)
 8.0
 	\end{lstlisting}
-\end{minipage}
+% \end{minipage}
