### 2021-11-01 #97 @matbesancon
when you do some modifications, you can rerun compilation yourself with `@whedon generate pdf`
