+++ paper/header.tex
@@ -5,7 +5,7 @@
 \author[1]{Marcelo Forets}
 \author[2]{Christian Schilling}
 \affil[1]{Universidad de la Rep\'{u}blica, Uruguay}
-\affil[2]{University of Konstanz, Germany}
+\affil[2]{Aalborg University, Denmark}
 
 \keywords{Set propagation, Geometry, Reachability analysis, Hybrid system, Support function, Formal verification}
 
