+++ paper/paper.yml
@@ -16,7 +16,7 @@ authors:
 affiliations:
   - name: Universidad de la Rep\'{u}blica, Uruguay
     index: 1
-  - name: University of Konstanz, Germany
+  - name: Aalborg University, Denmark
     index: 2
 date: 10 September 2021
 bibliography: ref.bib
