+++ README.md
@@ -1,7 +1,7 @@
 # [LazySets.jl](https://github.com/JuliaReach/LazySets.jl): Scalable Symbolic-Numeric Set Computations
 
 by [Marcelo Forets](https://github.com/mforets) (Universidad de la República, Uruguay) and
-[Christian Schilling](https://github.com/schillic) (University of Konstanz, Germany)*
+[Christian Schilling](https://github.com/schillic) (Aalborg University, Denmark)*
 
 **Both authors contributed equally.*
 
