### 2021-11-01 #97 @mforets
> the section on a statement of need. This is fine if this doesn't have this name but there should be at least a paragraph in the paper stating it

The statement of need can be found in paragraph 2 of the Introduction (Section 1). We just made a small change describing who the target audience is.
