+++ paper/introduction.tex
@@ -20,6 +20,7 @@ LazySets is designed to work with very high-dimensional sets but also provides s
 %
 Finally, LazySets is well integrated with the Julia ecosystem for scientific computing~\cite{bezanson2017julia}, and additional functionality is available upon loading optional packages.
 %
+The target audience of LazySets are researchers who use symbolic and numeric set computations.
 
 \smallskip
 
@@ -44,3 +45,4 @@ The code to reproduce the figures can be found in Appendix~\ref{sec:code_example
 	\caption{Example of set propagation using LazySets.}
 	\label{fig:lotka_volterra}
 \end{figure}
+
