### 2021-11-02 #97 @blegat
The paper is well written, I would recommend acceptance with minor revision. My comments are the following:

* Page 1 line 13 of right column: "comment" -> "comment**s**"
* "Optimization over a polyhedron corresponds to solving a linear program" -> this assumes that the objective function is linear
* "parameters to represent" should probably be "parameters that represent" or "parameters needed to represent"
* “common functionality” -> “common functionalities” or “a common functionality”
* “fresh set” -> “new set” or “fresh new set”
* In the LaTeX formula at the beginning of the last paragraph of page 3, none of the symbols are defined, I would say that CH is defined in Table 1, X0, E_+ are sets and Phi is a matrix (as defined later).
* In the definition of the support function, “maximum” should be supremum unless necessary assumptions are met for X
* “over polygon” -> “over the polygon” or “over a polygon”
* “numeric queries”: what do you mean ? Do you mean “numerical queries” ?
* Section 4.2, “Fig. 2 (left)” should be “Fig. 2 (right)”
* In Section 5.1, the first equation does not render well. It should be either in LaTeX math or in monospace font (e.g. wth texttt)
* In caption of Fig. 5, the first “(right)” should be “(left)”
* In the second line of the right column of page 10, “d1” and “d2” are not define. Moreover, d should have length 2n in this equation while it is defined to have length n
* The reference to MathOptInterface should be update: it is now published and contains a DOI (since a few days ago). “MathOptInterface” should be capitalized
* In the reference of TaylorModels.jl, “Taylormodels.jl” -> M should be capitalized
