+++ paper/introduction.tex
@@ -30,7 +30,7 @@ More advanced topics on type composition and conversion are introduced in sectio
 %
 Several applications are included in Section~\ref{sec:applications}.
 %
-We provide references and comment on related libraries in Section~\ref{sec:conclusion}.
+We comment on related libraries in Section~\ref{sec:conclusion}.
 %
 Installation instructions can be found in Appendix~\ref{sec:installation}.
 %
