+++ paper/basic.tex
@@ -88,7 +88,7 @@ HalfSpace([1.0, ..., 1.0], 10.0)
 
 
 The most widely used convex sets in various disciplines are (convex) \emph{polyhedra}, which are characterized as the finite intersection of half-spaces.
-Optimization over a polyhedron corresponds to solving a linear program.
+Optimization over a polyhedron with linear objective corresponds to solving a linear program.
 In Fig.~\ref{fig:supfunc} we show an example of a \emph{polytope} (a bounded polyhedron) with seven (linear) constraints in orange and a half-space in blue.
 
 \smallskip
