+++ paper/basic.tex
@@ -189,7 +189,7 @@ We are working toward having non-convex sets, such as set unions, in the same ty
 
 We have seen that we can interact with sets by checking properties.
 Importantly, we can also apply set operations to sets for constructing new sets.
-(By default the result is a fresh set and the original set is not manipulated.)
+(By default the result is a new set instance and the original set instance is not manipulated.)
 For example, one common set operation is to translate (or shift) every element in the set by a constant vector.
 
 \begin{minipage}{\linewidth}
