### 2021-11-03 9aa0837
explain symbols
