+++ paper/applications.tex
@@ -37,7 +37,7 @@ While writing this article, we received the following question:
 We need to plot polyhedra given in a form like this:
 
 \begin{center}
-	2*x1 >= 0 \& 3*x2+1.7*x3 >=0
+	\texttt{2 * x1 >= 0 \& 3 * x2 + 1.7 * x3 >= 0}
 \end{center}
 
 Is there a way to plot a 2D projection with LazySets?
