+++ paper/applications.tex
@@ -158,7 +158,7 @@ overapproximate(vTM, Hyperrectangle)
 	\centering
 	\includegraphics[height=3.5cm, keepaspectratio]{img/taylormodel_linear}
 	\includegraphics[height=3.5cm, keepaspectratio]{img/taylormodel_nonlinear}
-	\caption{Exact conversion of a linear Taylor model to a zonotope against an inexact evaluation using interval arithmetic (right). Approximate conversion of a non-linear Taylor model using a single zonotope, and range evaluation with interval arithmetic by splitting its domain into $100$ intervals (right).}
+	\caption{Exact conversion of a linear Taylor model to a zonotope against an inexact evaluation using interval arithmetic (left). Approximate conversion of a non-linear Taylor model using a single zonotope, and range evaluation with interval arithmetic by splitting its domain into $100$ intervals (right).}
 	\label{fig:taylormodelsconversion}
 \end{figure}
 
