### 2021-11-03 935c562
fix support function of Cartesian product
