+++ paper/macros.tex
@@ -3,4 +3,5 @@
 \newcommand{\R}{\ensuremath{\mathbb{R}}\xspace}
 \newcommand{\X}{\ensuremath{X}\xspace}
 \newcommand{\Y}{\ensuremath{Y}\xspace}
+\newcommand{\Z}{\ensuremath{Z}\xspace}
 \newcommand{\CH}{\ensuremath{\mathit{CH}}\xspace}
