+++ paper/ref.bib
@@ -78,16 +78,6 @@
 	doi          = {10.5281/zenodo.4993670},
 }
 
-@misc{legat2020mathoptinterface,
-	title = {MathOptInterface: a data structure for mathematical optimization problems},
-	author = {Beno{\^i}t Legat and Oscar Dowson and Joaquim Dias Garcia and Miles Lubin},
-	year = {2020},
-	eprint = {2002.03447},
-	archivePrefix = {arXiv},
-	primaryClass = {math.OC},
-	url = {https://arxiv.org/abs/2002.03447},
-}
-
 @article{DunningHuchetteLubin2017,
 	author = {Iain Dunning and Joey Huchette and Miles Lubin},
 	title = {{JuMP}: {A} {M}odeling {L}anguage for {M}athematical {O}ptimization},
@@ -101,10 +91,11 @@
 
 @article{MathOptInterface-2021,
 	title={{M}ath{O}pt{I}nterface: a data structure for mathematical optimization problems},
-	author={Legat, Benoit and Dowson, Oscar and Garcia, Joaquim and Lubin, Miles},
+	author={Legat, Beno{\^i}t and Dowson, Oscar and Dias Garcia, Joaquim and Lubin, Miles},
 	journal={INFORMS Journal on Computing},
-	year={in press},
-	publisher={INFORMS}
+	year={2021},
+	publisher={INFORMS},
+	doi={10.1287/ijoc.2021.1067}
 }
 
 @misc{sidrane2021overt,
