+++ paper/appendix.tex
@@ -5,6 +5,7 @@ To use LazySets, first install Julia version v1.3 or higher\footnote{Julia binar
 followed by
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 pkg> add LazySets
 \end{lstlisting}
@@ -12,6 +13,7 @@ pkg> add LazySets
 To load the package in a Julia session, do \code{using}, e.g.
 
 \begin{minipage}{\linewidth}
+	\vspace{-\abovedisplayskip}
 	\begin{lstlisting}
 julia> using LazySets
 
@@ -102,6 +104,7 @@ The set $\X_0$ is a ball in the infinity norm of radius $0.1$ centered in $[1, 0
 In the context of reachability analysis for linear differential equations \cite{BogomolovFFVPS18}, the set $\X_0$ corresponds to the initial states, $E_+$ accouts for bloating terms, and $\Phi = e^{A\delta}$ is the state-transition matrix for some matrix $A$ and time step $\delta > 0$.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 A = [0 1; -(4π)^2 0]
 X₀ = BallInf([1.0, 0.0], 0.1)
@@ -122,6 +125,7 @@ E₊ = Hyperrectangle(zeros(2), r)
 In Fig.~\ref{fig:supfunc}, the set $\X$ is a polygon in vertex representation. Such a \code{VPolygon} can be constructed from a vector of points, or simply a matrix where each column corresponds to the coordinates of a point. (For higher-dimensional sets in vertex representation the set type \code{VPolytope} is used).
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 # two-dimensional polygon in vertex representation
 X = VPolygon([-3   -2   0   1  2  0  -0.8;
@@ -132,6 +136,7 @@ X = VPolygon([-3   -2   0   1  2  0  -0.8;
 The supporting half-space of $\X$ is computed by evaluation of the support function along the direction of interest.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 # computing a supporting half-space
 d = [-1.0, 1.0]
@@ -147,6 +152,7 @@ The set $\X$ is the same as in Fig.~\ref{fig:supfunc}.
 We overapproximate it with a box (\code{Y}) and two zonotopes (\code{Z} and \code{W}).
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 Y = box_approximation(X)
 Z = overapproximate(X, Zonotope, PolarDirections(3))
@@ -158,6 +164,7 @@ The third argument to \code{overapproximate} here represents a list of vectors t
 The type \code{PolarDirections} lazily represents vectors that uniformly cover the unit disc, starting with the vector $(1, 0)^T$.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 collect(PolarDirections(5))
 5-element Vector{Vector{Float64}}:
@@ -176,6 +183,7 @@ The linear Taylor models are $p_1(t) = 2 + t$ and $p_2(t) = 0.9 + 3t$ with remai
 The non-linear case has $p_1(t)$ as in the linear one and $p_2(t) = 0.9 + 3t + 3t^2$ with the same remainders and domain as in the linear case.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 using TaylorModels
[…]
