+++ paper/basic.tex
@@ -66,6 +66,7 @@ which consists of all points on one side of the corresponding hyperplane.
 In LazySets these sets are constructed from $a$ and $b$. For example, the two-dimensional hyperplane $x = 1$ (resp. the half-space $x \leq 1$) are:
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 julia> a = [1.0, 0.0]; b = 1.0
 
@@ -76,8 +77,11 @@ julia> HalfSpace(a, b)
 HalfSpace{Float64,Vector{Float64}}([1.0, 0.0], 1.0)
 \end{lstlisting}
 \end{minipage}
+%
 Higher-dimensional sets are defined in a similar fashion; for instance, the 100-dimensional half-space $x_1 + \ldots + x_{100} \leq 10$ is:
+%
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 	\begin{lstlisting}
 julia> a = fill(1.0, 100); b = 10.0
 
@@ -102,6 +106,7 @@ For instance, the \code{BallInf} represents a hypercube specified by the center 
 In $n$ dimensions, a hypercube has $2^n$ vertices, but creating an $1{,}000$-dimensional \code{BallInf} is instantaneous.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 julia> @time BallInf(zeros(1000), 1.0)
 0.000005 seconds (2 allocations: 7.969 KiB)
@@ -117,6 +122,7 @@ A general approach to do that is rejection sampling, which picks a random point 
 We can thus use rejection sampling with any set type that implements a membership test.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 julia> ones(1000) ∈ BallInf(zeros(1000), 1.0)
 true
@@ -132,6 +138,7 @@ For example, we can ask for the list of vertices of a polytope.
 We have seen that a hypercube is represented by the center and the radius, so the vertices need to be computed on demand.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 julia> vertices_list(BallInf([1.0, 4.0], 1.0))
 4-element Vector{Vector{Float64}}:
@@ -145,7 +152,8 @@ julia> vertices_list(BallInf([1.0, 4.0], 1.0))
 Equality of sets in the mathematical sense can be checked via \code{isequivalent} (which by default checks mutual inclusion):
 
 \begin{minipage}{\linewidth}
-	\begin{lstlisting}
+\vspace{-\abovedisplayskip}
+\begin{lstlisting}
 julia> X = Interval(-1, 1) × Interval(-1, 1)
 CartesianProduct{Float64,
   Interval{...}, Interval{...}}(...)
@@ -193,6 +201,7 @@ Importantly, we can also apply set operations to sets for constructing new sets.
 For example, one common set operation is to translate (or shift) every element in the set by a constant vector.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 julia> B1 = BallInf([1.5, 2.0], 1.0)
 julia> B2 = translate(B1, [1.5, -1.0])
@@ -208,6 +217,7 @@ For most operations this is generally not the case.
 For instance, the intersection of two half-spaces is itself not a half-space but a polyhedron.
 
 \begin{minipage}{\linewidth}
+\vspace{-\abovedisplayskip}
 \begin{lstlisting}
 # intersect {x | x <= 1} and {x | x >= 0}
 julia> P = intersection(HalfSpace([1.0], 1.0),

[…]
