### 2021-11-03 #97 @schillic
@blegat Thanks for the thorough reading! We added https://github.com/JuliaReach/LazySets-JuliaCon21/pull/2 to address the review. Feel free to comment there.
