### 2021-11-04 #97 @gdalle
## General review

This paper introduces a very rich and thoughtfully-crafted package for set representation. The key features of LazySets.jl are clearly stated and illustrated with very visual examples. Its natural syntax, high performance, integration into the Julia ecosystem and extensive documentation are tremendous assets. **I recommend acceptance with minor revision.**

To make the paper even better, the authors could emphasize why the core functionalities of their package (like support function computations) are precisely the ones we need, especially in fields like convex analysis or reachability. A more thorough comparison with competing libraries would also be welcome. Provided these two points are addressed, I will be able to check the remaining boxes in the Context section below.

For some reason I am not able to edit the checklist, so I reproduce it below.

## Reviewer checklist

- [x] As the reviewer I confirm that I have read the JuliaCon conflict of interest policy and that there are no conflicts of interest for me to review this work.

- [x] I confirm that I read and will adhere to the JuliaCon code of conduct.

- [ ] Repository: Is the source code for this software available at the [repository url](https://github.com/JuliaReach/LazySets-JuliaCon21)?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [x] Authorship: Has the submitting author (@mforets) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

- [x] Authors: Does the paper.tex file include a list of authors with their affiliations?
- [x] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
- [x] Page limit: Is the page limit for full papers respected by the submitted document?

- [ ] Context: is the scientific context motivating the work correctly presented?
- [x] Methodology: is the approach taken in the work justified, presented with enough details and reference to reproduce it?
- [ ] Results: are the results presented and compared to approaches with similar goals?
