### 2021-11-04 JuliaReach/LazySets-JuliaCon21#4 @gdalle
Review 2
Hello, and congratulations on your exceptional work!

Here are some line-by-line suggestions relative to your paper. Regarding the package itself, I don't have much to say, except that some tests are throwing warnings about polyhedra / LP solvers, is that expected behavior?

**General**

Some references dont display DOIs (if they don't have any, maybe the Arxiv reference would be relevant?): 4, 11, 23. I didn't manage to find reference 21 with a quick Google search.

It is probably due to the template, but copy-pasting code from the PDF adds unwanted spaces between every pair of characters.

**Page 1**

> For a visual example showing the sets of states reachable by the Lotka-Volterra equations, see Fig. 1.

Additional comments / legend / axis labels on Figure 1 would be welcome: a reader not familiar with reachability analysis might struggle to see the connection with Lotka-Volterra equations.

**Page 2**

> `HalfSpace([1.0,...,1.0],10.0)`

In the second code snippet, why is `HalfSpace` no longer displayed as a parametric type, as in `HalfSpace{Float64,Vector{Float64}}`?

> The most widely used convex sets in various disciplines are (convex) polyhedra

It would be nice to justify the omnipresence of linear programs: they are able to model a wide variety of real-life optimization problems

> "In Fig. 2 we show an example of a polytope (a bounded polyhedron) with seven (linear) constraints in orange and a half-space in blue."

The previous sentence sounds like the half-space is part of the polytope somehow. Maybe add a comma after "orange"?

> Other typical properties that can be checked for sets X and Y are emptiness, inclusion, and having no point in common

What is the complexity of these operations? They seem much more involved than membership checking.

**Page 3**

> For a complete list of the set operations available in LazySets we refer to Table 1.

The distinction between lazy constructor and concrete function in Table 1 is not obvious at this stage of the paper: this would deserve an small explanation or a reference to what comes next.

> To give an example, the Minkowski sum of a square and a disc (two-dimensional balls in the infinity norm and Euclidean norm) is not representable with a basic type in LazySets. Hence `minkowski_sum` will yield an error. But we can apply the lazy `MinkowskiSum` operation, which itself subtypes LazySet.

Here it would help to repeat what `MinkowskiSum` is, as opposed to `minkowski_sum`, or refer to Table 1.

> (This also motivates the name of LazySets)

That remark may come earlier, maybe even in the introduction.

> $X_0$ and $E_+$ are sets

Why the strange notation $E_+$? Why not $X$ and $Y$? If it has something to do with reachability applications, it would be nice to know just a little about the context.

**Page 4**

> A standard approach to operate with compact and convex sets is to use the support function

Could you say what this function is used for? From what I understand, its computation is one of the main selling points of the package, yet the reader unfamiliar with convex analysis may not see why it is so important

> LazySets offers `ρ(d, X)`

Why does $X$ have a different font than \texttt?

> Fig. 2 (left) illustrates the evaluation of the support function over the polygon $X$

Wouldn't it make more sense to plot the vector $d$ as going from $(0, 0)$ to $(-3, 0.6)$?

> its support function can be computed by solving a linear program (LP)

Linear programs were introduced earlier but the acronym was not defined

> when we make queries to the lazy set such as asking for its support function along $d = (−1,1)^T$

You previously wrote this vector as $d = (-1.0, 1.0)^T$ (with decimals)

> `julia> @btime ρ($d,$Ω0)`

What does benchmarking demonstrate here?

> For many applications it is sufficient to evaluate the support function in only a few directions.

Again, it would really add a lot of value to list some of these applications

> To still give mathematical guarantees, one usually aims for either over- or underapproximations.

Underapproximations are not mentioned in what follows, are they not supported?

**Page 5**

> Fig. 3: Left picture: Overapproximation of the polytope from Fig. 2 (purple)

Keeping the polytope orange as before would be lovely if it were possible

> In two dimensions (or the projection of a higher-dimensional set into two dimensions), LazySets can compute $\varepsilon$-close over-approximations

Why isn't it possible in higher dimensions?

> Such overapproximations generally yield an `HPolytope` if X is bounded

Types from Polyhedra.jl are used or mentioned several times without being formally introduced. A sentence about H vs V-representation may clarify things

> The Makie plot and the two-dimensional projection are shown in Fig. 4

The 3D Makie plot is a bit too dark to see inside the volume

**Page 6**

> Using ForwardDiff.jl to get the gradient of [...] throws the following error:

I think this may not be very clear to people unfamiliar with autodiff, especially the ForwardDiff.jl package. A motivated use of automatic differentiation would be more illuminating

> Taylor models are, informally, sets defined by the image of a polynomial map with interval remainders

This sentence is unclear to me, and this may be why I struggle to understand the rest of the example

> `# with a zonotopes`

Typo (no plural)

> Fig. 5

A legend highlighting the difference between $Z$ and $H$ on both sides would help

> queries about the zonotope provide quick information about the exact set, more accurate than the box approximation:

Would it be possible to plot both vectors $d$ on Fig 5?

> a comparison against an evaluation using interval arithmetic with many small boxes

This formulation seems vague, I'm not sure I understand what you do there

> JuliaReach builds on sound scientific approaches

Such as?

> The model has distributed initial conditions for all variables

What does this mean?

**Page 7**

> Fig. 6

The legend is not quite enough for me to understand what is going on here. I struggle to see what happens in the case with non-deterministic inputs

> initial conditions are defined as a flat hyperrectangle of radius 0.1 along the x coordinate:

The following code snippet doesn't seem necessary to me

**Page 8**

> Moreover, linear systems can be embedded in algorithms to analyze nonlinear systems

This helps justify the focus on polyhedra / linear constraints

> the approach is able to handle a large number of sets (100 million sets in zonotope representation), and it was shown to be an order of magnitude faster than competing tools [2].

This kind of quantitative performance evaluation is very important, I would underline it more. The same goes for the ARCH-COMP win, if you can elaborate on that.

> 6.2 Related libraries

I find it strange that you wait until the conclusion to mention competing tools. I think at least a brief overview should come sooner, and most importantly you should say what these competitors are lacking in functionality, performance, documentation or whatever other aspects distinguishes LazySets from the pack. You could also add links to repos for these packages, as you did with the Julia ones.

> In our experience, one can often obtain speed-ups within several orders of magnitude by adding new methods.

If the reader isn't familiar with multiple dispatch, this may not be clear. Give an example you encountered during development?

**Page 10**

> C. Code used in examples

I haven't found the code for the examples in the paper repo, do you plan to add it soon?
