### 2021-11-04 #97 @gdalle
I have put more detailed comments in this issue https://github.com/JuliaReach/LazySets-JuliaCon21/issues/4
