### 2021-11-10 #97 @gdalle
I am waiting on a revised version by @mforets and @schillic to check the remaining boxes
