+++ paper/ref.bib
@@ -3,7 +3,8 @@
 	author={Skogestad, Sigurd and Postlethwaite, Ian},
 	volume={2},
 	year={2007},
-	publisher={John Wiley \& Sons}
+	publisher={John Wiley \& Sons},
+	isbn = {0470011688}
 }
 
 @inproceedings{ARCHCOMP20NONLINEAR,
@@ -99,12 +100,19 @@
 }
 
 @misc{sidrane2021overt,
-	title={{OVERT}: An Algorithm for Safety Verification of Neural Network Control Policies for Nonlinear Systems}, 
-	author={Chelsea Sidrane and Amir Maleki and Ahmed Irfan and Mykel J. Kochenderfer},
-	year={2021},
-	eprint={2108.01220},
-	archivePrefix={arXiv},
-	primaryClass={cs.LG}
+  author    = {Chelsea Sidrane and
+               Amir Maleki and
+               Ahmed Irfan and
+               Mykel J. Kochenderfer},
+  title     = {{OVERT:} An Algorithm for Safety Verification of Neural Network Control
+               Policies for Nonlinear Systems},
+  journal   = {CoRR},
+  volume    = {abs/2108.01220},
+  year      = {2021},
+  url       = {https://arxiv.org/abs/2108.01220},
+  howpublished = {\url{https://arxiv.org/abs/2108.01220}},
+  eprinttype = {arXiv},
+  eprint    = {2108.01220}
 }
 
 @inproceedings{soto2021synthesis,
@@ -132,7 +140,7 @@
 }
 
 @article{schilling2021verification,
-	title={Verification of Neural-Network Control Systems by Integrating Taylor Models and Zonotopes
+	title={Verification of Neural-Network Control Systems by Integrating {T}aylor Models and Zonotopes
 	},
 	author={Schilling, Christian and Forets, Marcelo and Guadalupe, Sebastian},
 	journal={Submitted.},
@@ -140,13 +148,14 @@
 }
 
 @article{forets2021combining,
-	title={Combining Set Propagation with Finite Element Methods for Time Integration in Transient Solid Mechanics Problems},
+	title={Combining set propagation with finite element methods for time integration in transient solid mechanics problems},
 	author={Forets, Marcelo and Caporale, Daniel Freire and Zerpa, Jorge M P{\'e}rez},
-	journal={CoRR},
-	year={2021},
-	url       = {https://arxiv.org/abs/2105.05841},
-	eprinttype = {arXiv},
-	eprint    = {2105.05841}
+	journal={Computers \& Structures},
+	year={2022},
+	volume = {259},
+	pages = {106699},
+	doi = {10.1016/j.compstruc.2021.106699},
+	url = {https://www.sciencedirect.com/science/article/pii/S0045794921002212},
 }
 
 @inproceedings{forets2020efficient,
@@ -231,12 +240,13 @@
   howpublished = {\url{https://github.com/JuliaIntervals/TaylorModels.jl}}
 }
 
-@inproceedings{BenetFSS19,
+@misc{BenetFSS19,
   author = {Benet, Luis and Forets, Marcelo and Sanders, David P. and Schilling, Christian},
   title = {{TaylorModels}.jl: {T}aylor models in Julia and its application to validated solutions of {ODE}s},
   booktitle = {{SWIM}},
   year = {2019},
-  url = {https://swim2019.ensta-paristech.fr/}
+  url = {https://swim2019.ensta-paristech.fr/},
+  howpublished = {\url{https://swim2019.ensta-paristech.fr/}}
 }
 
 @inproceedings{Althoff15,
