+++ paper/ref.bib
@@ -334,3 +334,18 @@
 	url       = {https://doi.org/10.1109/TCAD.2020.3012859},
 	doi       = {10.1109/TCAD.2020.3012859}
 }
+
+@book{kochenderfer2019algorithms,
+  title={Algorithms for optimization},
+  author={Kochenderfer, Mykel J and Wheeler, Tim A},
+  year={2019},
+  publisher={Mit Press}
+}
+
+@book{dantzig1998linear,
+  title={Linear Programming and Extensions},
+  author={Dantzig, George Bernard},
+  volume={48},
+  year={1998},
+  publisher={Princeton University Press}
+}
