+++ paper/basic.tex
@@ -1,5 +1,5 @@
 \begin{table*}[tb]
-	\tbl{Set operations available in LazySets. The result of the last three operations is generally not convex even if used with convex operands. For binary operations (marked with \binary{$\cdot$}) there is also an $n$-ary lazy version with the suffix \code{Array}, e.g., \code{MinkowskiSumArray}. Unicode symbols (as mentioned in the column ``Short form'') are entered in the Julia REPL by typing the \LaTeX\ command (e.g.: \code{\textbackslash{}oplus} for $\oplus$) followed by pressing the ``Tab'' key. See Appendix~\ref{sec:setops} for some central definitions.}{
+	\tbl{Set operations available in LazySets. The distinction between lazy and concrete functions is explained in Section~\ref{sec:lazy}. The result of the last three operations is generally not convex even if used with convex operands. For binary operations (marked with \binary{$\cdot$}) there is also an $n$-ary lazy version with the suffix \code{Array}, e.g., \code{MinkowskiSumArray}. Unicode symbols (as mentioned in the column ``Short form'') are entered in the Julia REPL by typing the \LaTeX\ command (e.g.: \code{\textbackslash{}oplus} for $\oplus$) followed by pressing the ``Tab'' key. See Appendix~\ref{sec:setops} for some central definitions.}{
 	\begin{tabular}{l l l l l}
 		\toprule
 		Operation name & Math form & Lazy function (constructor) & Short form & Concrete function \\
@@ -78,7 +78,7 @@ HalfSpace{Float64,Vector{Float64}}([1.0, 0.0], 1.0)
 \end{lstlisting}
 \end{minipage}
 %
-Higher-dimensional sets are defined in a similar fashion; for instance, the 100-dimensional half-space $x_1 + \ldots + x_{100} \leq 10$ is:
+Higher-dimensional sets are defined in a similar fashion; for instance, the 100-dimensional half-space $x_1 + \ldots + x_{100} \leq 10$ is:\footnote{Sometimes we abbreviate the output to improve readability.}
 %
 \begin{minipage}{\linewidth}
 \vspace{-\abovedisplayskip}
@@ -93,7 +93,8 @@ HalfSpace([1.0, ..., 1.0], 10.0)
 
 The most widely used convex sets in various disciplines are (convex) \emph{polyhedra}, which are characterized as the finite intersection of half-spaces.
 Optimization over a polyhedron with linear objective corresponds to solving a linear program.
-In Fig.~\ref{fig:supfunc} we show an example of a \emph{polytope} (a bounded polyhedron) with seven (linear) constraints in orange and a half-space in blue.
+Linear programs can model a wide variety of real-life optimization problems and can be solved efficiently \cite{dantzig1998linear,kochenderfer2019algorithms}.
+On the left of Fig.~\ref{fig:supfunc} we show an example of a \emph{polytope} (a bounded polyhedron) in orange, with seven (linear) constraints, and a half-space in blue.
 
 \smallskip
 
@@ -130,6 +131,10 @@ true
 \end{minipage}
 
 Other typical properties that can be checked for sets $\X$ and $\Y$ are emptiness ($\X = \emptyset$; \code{isempty}), inclusion ($\X \subseteq \Y$; \code{issubset}), and having no point in common ($\X \cap \Y = \emptyset$; \code{isdisjoint}).
+The complexity of such operations depends on the representation of the set.
+Take for example a polyhedron represented by the list of its linear constraints.
+Emptiness can be checked by formulating a feasibility linear program.
+Inclusion and disjointness can be checked using the support function (see Section~\ref{sec:supfun}).
 
 \smallskip
 
@@ -198,6 +203,7 @@ We are working toward having non-convex sets, such as set unions, in the same ty
 We have seen that we can interact with sets by checking properties.
[…]
