+++ paper/img/plots.jl
@@ -22,13 +22,16 @@ xticklatex(vec, digits) = (vec, __toL.(vec, Ref(digits)))
 
 X = VPolygon([[0.8, 2.2], [2.1, 1], [1.9, -1], [0, -2],
               [-2, -2], [-3, 0.6], [-0.8, 1.8], [0.8, 2.2]])
-d = [-1.0, 1.0]
+d = [-1, 1]
 
 sv = @show σ(d, X)
 
 sf = @show ρ(d, X)
 
 H = HalfSpace(d, sf)
+
+dn = normalize(d)
+sfn = ρ(dn, X)
 
 fig = p(xlim=(-4.1, 2.5), ylim=(-2.5, 2.5), ratio=1, xlab=L"x_1", ylab=L"x_2",
            legendfontsize=25,
@@ -43,9 +46,10 @@ fig = p(xlim=(-4.1, 2.5), ylim=(-2.5, 2.5), ratio=1, xlab=L"x_1", ylab=L"x_2",
 p!(fig, H, alpha=0.5, color=:lightblue, lw=3.0, ls=:dash, lc=:black)
 p!(fig, X, color=:orange, lab="")
 #p!(fig, Singleton([0.0, 0.0]), color=:black, lab="")
-p!(fig, [sv[1], sv[1] + d[1]], [sv[2], sv[2] + d[2]], linecolor=:red, arrow=:arrow,
-   linestyle=:dot, width=6, annotations=(-3.3, 1.5, text(L"d", 28)))
-p!(fig, Singleton(sv), alpha=1, markersize=10, color=:green, lab="")
+p!(fig, [0, dn[1]*sfn], [0, dn[2]*sfn], linecolor=:red, arrow=:both,
+   linestyle=:dot, width=4, annotations=(0.6, 1.4, text(L"\rho(d\prime, X)", 28)))
+p!(fig, Singleton(sv), alpha=1, markersize=10, color=:green, lab="",
+   annotations=(-2.5, 0, text(L"\sigma(d, X)", 28)))
 savefig(fig, "supfunc.pdf")
 
 fig = p(xlim=(-4.1, 2.5), ylim=(-2.5, 2.5), ratio=1, xlab=L"x_1", ylab=L"x_2",
@@ -122,7 +126,7 @@ plot!(fig, map(p -> LineSegment(zeros(2), p), PolarDirections(3)), lw=5.,
 savefig(fig, "polardirs.pdf")
 
 # --------------------------------------------
-# Lazy set Ω₀ = CH(X₀, Φ*X₀ ⊕ E₊) example
+# Lazy set Ω₀ = CH(X₀, Φ*X₀ ⊕ Y) example
 # --------------------------------------------
 
 A = [0 1; -(4π)^2 0]
@@ -131,9 +135,9 @@ X₀ = BallInf([1.0, 0.0], 0.1)
 Φ = exp(A*δ)
 
 r = [0.05477208, 0.07676220]
-E₊ = Hyperrectangle(zeros(2), r)
+Y = Hyperrectangle(zeros(2), r)
 
-Ω₀ = CH(X₀, Φ*X₀ ⊕ E₊)
+Ω₀ = CH(X₀, Φ*X₀ ⊕ Y)
 
 # make boxes with rounded corners? (doesn't work)
 # using TreeView: TikzGraphs # required to change plot aspects
