+++ paper/lazy.tex
@@ -63,18 +63,26 @@ A standard approach to operate with compact and convex sets $\X \in \R^n$ is to 
 %
 The support function along direction $d \in \R^n$, denoted $\rho(d, \X)$, is the maximum of $d^T x$ over all $x \in \X$, i.e., the support function describes the (signed) distance of the supporting hyperplane in direction $d$ from the origin.
 %
+It can be used to efficiently find the boundary of a set in a given direction.
+%
 The maximizers are called \emph{support vectors} $\sigma(d, \X)$. Intuitively, the support vectors are the extreme points of $\X$ in direction $d$.
 %
 Basic properties of the support function are given in Appendix~\ref{sec:supfunc_properties}.
 
+For various set representations, the support function is known analytically and can be efficiently evaluated numerically. Such cases include hyperrectangular sets and zonotopic sets.
+%
+For sets with less structure, e.g., if $\X$ is a polytope in half-space representation, its support function can be computed by solving a linear program, for which fast and robust solvers exist.
+%
+But the main advantage of using the support function in LazySets lies in the extensive use of \emph{composition rules}, as we describe later.
+
 \smallskip
 
-LazySets offers \code{$\uprho$(d, \X)} (or \code{support\_function(d, \X)}) to compute the support function $\rho(d,\X)$, and \code{$\upsigma$(d, \X)} (or \code{support\_vector(d, \X)}) to compute (some) support vector $\sigma(d,\X)$. Fig.~\ref{fig:supfunc} (left) illustrates the evaluation of the support function over the polygon $\X$ (orange) along direction $(-1.0, 1.0)^T$.
+LazySets offers \code{$\uprho$(d, X)} (or \code{support\_function(d, X)}) to compute the support function $\rho(d,\X)$, and \code{$\upsigma$(d, X)} (or \code{support\_vector(d, X)}) to compute (some) support vector $\sigma(d,\X)$. Fig.~\ref{fig:supfunc} (left) illustrates the evaluation of the support function over the polygon $\X$ (orange) along direction $(-1, 1)^T$.
 
 \begin{minipage}{\linewidth}
 \vspace{-\abovedisplayskip}
 \begin{lstlisting}
-julia> d = [-1.0, 1.0]
+julia> d = [-1, 1]
 
 julia> ρ(d, X) # or support_function(d, X)
 3.6
@@ -89,34 +97,35 @@ julia> σ(d, X) # or support_vector(d, X)
 	\includegraphics[width=0.49\linewidth, keepaspectratio]{img/supfunc}
 	\includegraphics[width=0.49\linewidth, keepaspectratio]{img/supfunc_oct}
 	\vspace*{1mm}
-	\caption{Supporting hyperplane of the set $\X$ along direction $d$ (left), and outer approximation of $\X$ using the eight directions of a regular octagon (right).}
+	\caption{Left: The supporting hyperplane of the set $\X$ along direction $d$.
+	In red we plot the distance of the hyperplane to the origin, which is given by $\rho(d', \X)$ where $d' = d / \Vert d \Vert$.
+	Right: An outer approximation of $\X$ using the eight directions of a regular octagon.}
 	\label{fig:supfunc}
 \end{figure}
-
-For various set representations, the support function is known analytically and can be efficiently evaluated numerically. Such cases include hyperrectangular sets and zonotopic sets.
-%
-For sets with less structure, e.g., if $\X$ is a polytope in half-space representation, its support function can be computed by solving a linear program (LP), for which fast and robust solvers exist.
-%
-But the main advantage of using the support function in LazySets lies in the extensive use of \emph{composition rules} as we describe next.
-
-\smallskip
 
[…]
