### 2021-11-11 f0c228e
better arrows in plot
