+++ paper/applications.tex
@@ -152,11 +152,11 @@ using TaylorModels
 
 # approximate a vector of Taylor models
 # with a zonotope
-overapproximate(vTM, Zonotope)
+Z = overapproximate(vTM, Zonotope)
 
 # approximate a vector of Taylor models
 # with a hyperrectangle
-overapproximate(vTM, Hyperrectangle)
+H = overapproximate(vTM, Hyperrectangle)
 	\end{lstlisting}
 \end{minipage}
 
@@ -164,39 +164,41 @@ overapproximate(vTM, Hyperrectangle)
 	\centering
 	\includegraphics[height=3.5cm, keepaspectratio]{img/taylormodel_linear}
 	\includegraphics[height=3.5cm, keepaspectratio]{img/taylormodel_nonlinear}
-	\caption{Left: Exact conversion of a linear Taylor model (red) to a zonotope (green) against an inexact evaluation using interval arithmetic (blue). Right: Approximate conversion of a non-linear Taylor model (red) using a single zonotope (green).}
+	\caption{Left: Exact conversion of a linear Taylor model (red) to a zonotope (green) against an inexact evaluation using interval arithmetic (blue). Right: Approximate conversion of a non-linear Taylor model (red) using a zonotope (green).}
 	\label{fig:taylormodelsconversion}
 \end{figure}
 
 In the case when the Taylor model is linear, the conversion is exact, because the zonotope stores linear dependencies between variables.
 %
-Hence, as is shown in Fig.~\ref{fig:taylormodelsconversion} (left), the zonotope approximation (\code{Z}) is better than directly using interval arithmetic to evaluate the range (\code{H}), and queries about the zonotope provide quick information about the exact set, more accurate than the box approximation:
+Hence, as is shown in Fig.~\ref{fig:taylormodelsconversion} (left), the zonotope approximation (\code{Z}) is better than directly using interval arithmetic to evaluate the range (\code{H}), and queries about the zonotope provide quick information about the exact set, more accurate than the box approximation.
 
-\begin{minipage}{\linewidth}
-	\vspace{-\abovedisplayskip}
-	\begin{lstlisting}
-julia> d = [-0.35, 0.93]
- 
-julia> ρ(d, Z)
-3.217
-
-julia> ρ(d, H)
-4.617000000000001
-\end{lstlisting}
-\end{minipage}
-
-If the Taylor model contains nonlinear terms, the zonotope provides an enclosure; a comparison against an evaluation using interval arithmetic with many small boxes is shown in Fig.~\ref{fig:taylormodelsconversion} (right).
+% \begin{minipage}{\linewidth}
+% 	\vspace{-\abovedisplayskip}
+% 	\begin{lstlisting}
+% julia> d = [-0.35, 0.93]
+%  
+% julia> ρ(d, Z)
+% 3.217
+% 
+% julia> ρ(d, H)
+% 4.617000000000001
+% \end{lstlisting}
+% \end{minipage}
 
 \smallskip
 
-We also mention that LazySets can be combined with the powerful interval constraint programming approach by simply loading the optional dependency \href{https://github.com/JuliaIntervals/IntervalConstraintProgramming.jl}{IntervalConstraintProgramming.jl}.
+If the Taylor model contains nonlinear terms, the zonotope provides only an enclosure but can still be more precise than the box approximation; a comparison is shown in Fig.~\ref{fig:taylormodelsconversion} (right); we obtain a more precise approximation of the Taylor model by partitioning its domain and evaluating each resulting block with a smaller box.
+
+\smallskip
+
+We also mention that LazySets can be combined with an interval-constraint-programming approach by simply loading the optional dependency \href{https://github.com/JuliaIntervals/IntervalConstraintProgramming.jl}{IntervalConstraintProgramming.jl}.
 
 
[…]
