+++ paper/macros.tex
@@ -4,4 +4,5 @@
 \newcommand{\X}{\ensuremath{X}\xspace}
 \newcommand{\Y}{\ensuremath{Y}\xspace}
 \newcommand{\Z}{\ensuremath{Z}\xspace}
+\newcommand{\U}{\ensuremath{U}\xspace}
 \newcommand{\CH}{\ensuremath{\mathit{CH}}\xspace}
