### 2021-11-15 b1e4522
add discretization reference
