+++ paper/ref.bib
@@ -349,3 +349,15 @@
   year={1998},
   publisher={Princeton University Press}
 }
+
+@article{ForetsS21,
+  author    = {Marcelo Forets and
+               Christian Schilling},
+  title     = {Conservative Time Discretization: {A} Comparative Study},
+  journal   = {CoRR},
+  volume    = {abs/2111.01454},
+  year      = {2021},
+  url       = {https://arxiv.org/abs/2111.01454},
+  eprinttype = {arXiv},
+  eprint    = {2111.01454}
+}
