+++ paper/lazy.tex
@@ -16,7 +16,7 @@ We wrap the operands in a new object that, by definition, represents the result 
 
 \smallskip
 
-As an illustrative example, suppose that we are interested in the formula $\Omega_0 = \CH(\X_0, \Phi \X_0 \oplus Y)$, where $\CH$ and $\oplus$ are defined in Table~\ref{tab:operations}, $X_0$ and $Y$ are sets, and $\Phi$ is a matrix defined in Appendix~\ref{sec:omega0}. Such formulas are prevalent in reachability analysis of linear initial-value problems, or nonlinear ones after some form of conservative linearization; see for example \cite{althoff2020set} and references therein. Given the sets $\X_0$, $Y$, and the matrix $\Phi$, we can write:
+As an illustrative example, suppose that we are interested in the formula $\Omega_0 = \CH(\X_0, \Phi \X_0 \oplus Y)$, where $\CH$ and $\oplus$ are defined in Table~\ref{tab:operations}, $X_0$ and $Y$ are sets, and $\Phi$ is a matrix defined in Appendix~\ref{sec:omega0}. Such formulas are prevalent in reachability analysis of linear initial-value problems, or nonlinear ones after some form of conservative linearization; see for example \cite{althoff2020set,ForetsS21} and references therein. Given the sets $\X_0$, $Y$, and the matrix $\Phi$, we can write:
 
 \begin{minipage}{\linewidth}
 \begin{lstlisting}
