+++ paper/img/plots.jl
@@ -1,333 +1,11 @@
-# -------------------
-# Preamble
-# -------------------
+include("preamble.jl")
 
-using LazySets
-import Plots
-using Plots: plot, plot!, text, lens!, bbox, font, savefig
-using LaTeXStrings
-using Plots.PlotMeasures
+include("plot_supfunc.jl")
 
-# custom plot shorthands
-p(args...; kwargs...) = plot(xlims=(0, 4), ylims=(0, 4), ratio=1, lab="", xlab="x₁", ylab="x₂", args...; kwargs...)
-p!(args...; kwargs...) = plot!(lab="", args...; kwargs...);
+include("plot_overapproximate.jl")
 
-# utils for plots
-__toL(x, digits) = L"%$(round(x, digits=digits))"
-xticklatex(vec, digits) = (vec, __toL.(vec, Ref(digits)))
-Plots.GR.setarrowsize(3)  # larger arrow heads
+include("plot_polyhedron.jl")
 
-# -------------------
-# Suport function
-# -------------------
+include("plot_VTK.jl")
 
-X = VPolygon([[0.8, 2.2], [2.1, 1], [1.9, -1], [0, -2],
-              [-2, -2], [-3, 0.6], [-0.8, 1.8], [0.8, 2.2]])
-d = [-1.0, 1.0]
-
-sv = @show σ(d, X)
-
-sf = @show ρ(d, X)
-
-H = HalfSpace(d, sf)
-
-dn = normalize(d)
-sfn = ρ(dn, X)
-
-fig = p(xlim=(-4.1, 2.5), ylim=(-2.5, 2.5), ratio=1, xlab=L"x_1", ylab=L"x_2",
-           legendfontsize=25,
-           tickfont=font(25, "Times"),
-           guidefontsize=25,
-           xguidefont=font(25, "Times"),
-           yguidefont=font(25, "Times"),
-           xtick = ([-4, -2, 0, 2], [L"-4", L"-2", L"0", L"2"]),
-           ytick = ([-2, 0, 2], [L"-2", L"0", L"2"]),
-           bottom_margin=0mm, left_margin=0mm, right_margin=0mm, top_margin=0mm, size=(500, 390))
-
-p!(fig, H, alpha=0.5, color=:lightblue, lw=3.0, ls=:dash, lc=:black)
-p!(fig, X, color=:orange, lab="")
-#p!(fig, Singleton([0.0, 0.0]), color=:black, lab="")
-p!(fig, [0, dn[1]*sfn], [0, dn[2]*sfn], linecolor=:red, arrow=:both,
-   linestyle=:dot, width=4, annotations=(0.6, 1.4, text(L"\rho(d\prime, X)", 28)))
-p!(fig, Singleton(sv), alpha=1, markersize=10, color=:green, lab="",
-   annotations=(-2.5, 0, text(L"\sigma(d, X)", 28)))
-savefig(fig, "supfunc.pdf")
-
-fig = p(xlim=(-4.1, 2.5), ylim=(-2.5, 2.5), ratio=1, xlab=L"x_1", ylab=L"x_2",
-           legendfontsize=25,
-           tickfont=font(25, "Times"),
-           guidefontsize=25,
-           xguidefont=font(25, "Times"),
-           yguidefont=font(25, "Times"),
-           xtick = ([-4, -2, 0, 2], [L"-4", L"-2", L"0", L"2"]),
-           ytick = ([-2, 0, 2], [L"-2", L"0", L"2"]),
-           bottom_margin=0mm, left_margin=0mm, right_margin=0mm, top_margin=0mm, size=(500, 390))
-
-Xoct = overapproximate(X, OctDirections(2))
[…]
