### 2021-11-15 7700a5d
update polyhedron plot
