+++ paper/applications.tex
@@ -83,12 +83,12 @@ plot3d(R)  # plot in three dimensions
 \begin{figure}
 	\centering
 	\hfill
-	\includegraphics[width=0.49\linewidth, height=3cm, keepaspectratio,clip,trim=65mm 14mm 62mm 19mm]{img/polyhedron3D}
+	\includegraphics[width=0.49\linewidth, height=3cm, keepaspectratio,clip,trim=40mm 14mm 62mm 35mm]{img/polyhedron3D}
 	\hfill
 	\includegraphics[width=0.49\linewidth, height=3cm, keepaspectratio]{img/polyhedron2D}
 	\hfill\,
 	\vspace*{1mm}
-	\caption{Three-dimensional polyhedron plotted using Makie (left) and a two-dimensional projection obtained by Polyhedra with CDDLib (right).}
+	\caption{Three-dimensional polyhedron plotted using Makie (left) and a two-dimensional projection obtained by Polyhedra with CDDLib (right). Edges of a triangulation of the polyhedron are plotted in dashed lines (red).}
 	\label{fig:polyhedra}
 \end{figure}
 
