+++ paper/ref.bib
@@ -360,7 +360,8 @@
   title={Algorithms for optimization},
   author={Kochenderfer, Mykel J and Wheeler, Tim A},
   year={2019},
-  publisher={Mit Press}
+  publisher={Mit Press},
+  isbn = {9780262039420}
 }
 
 @book{dantzig1998linear,
@@ -368,7 +369,8 @@
   author={Dantzig, George Bernard},
   volume={48},
   year={1998},
-  publisher={Princeton University Press}
+  publisher={Princeton University Press},
+  doi = {10.1515/9781400884179}
 }
 
 @article{ForetsS21,
