+++ paper/acks.tex
@@ -1,7 +1,9 @@
 First, we thank all those of have contributed to LazySets\footnote{The full list of contributors is available \href{https://github.com/JuliaReach/LazySets.jl/graphs/contributors}{on Github}.}.
 %
-We are specially thankful to the following people for discussions and contributions at various stages of this work (sorted alphabetically): Tomer Arnon, Luis Benet, Aadesh Deshmukh, Goran Frehse, Daniel Freire, Bruno Garate, Ander Gray, Sebastian Guadalupe, Nikolaos Kekatos, Benoit Legat, Jorge P\'{e}rez Zerpa, Kostiantyn Potomkin, David P. Sanders, Frederic Viry, Ueli Wechsler and Peng Yu.
+We are specially thankful to the following people for discussions and contributions at various stages of this work (sorted alphabetically): Tomer Arnon, Luis Benet, Aadesh Deshmukh, Goran Frehse, Daniel Freire, Bruno Garate, Ander Gray, Sebastian Guadalupe, Nikolaos Kekatos, Beno\^it Legat, Jorge P\'{e}rez Zerpa, Kostiantyn Potomkin, David P. Sanders, Frederic Viry, Ueli Wechsler and Peng Yu.
 %
 Finally, we thank the people behind Julia Seasons of Contributions and Google Summer of Code programs for their financial support of several students to further develop LazySets.
 %
 The first author is partly supported by Agencia Nacional de Investigaci\'{o}n e Innovaci\'{o}n (ANII), Uruguay.
+%
+Finally, we greatly appreciate Beno\^it Legat and Guillaume Dalle for their work in revising this paper.
