+++ paper/ref.bib
@@ -1,3 +1,24 @@
+@book{featherstone2014rigid,
+	title={Rigid body dynamics algorithms},
+	author={Featherstone, Roy},
+	year={2014},
+	publisher={Springer},
+	isbn = {978-0-387-74314-1}
+}
+
+@article{BaydinPRS17,
+	author    = {Atilim Gunes Baydin and
+	Barak A. Pearlmutter and
+	Alexey Andreyevich Radul and
+	Jeffrey Mark Siskind},
+	title     = {Automatic Differentiation in Machine Learning: a Survey},
+	journal   = {J. Mach. Learn. Res.},
+	volume    = {18},
+	pages     = {153:1--153:43},
+	year      = {2017},
+	url       = {http://jmlr.org/papers/v18/17-468.html}
+}
+
 @book{skogestad2007multivariable,
 	title={{M}ultivariable feedback control: analysis and design},
 	author={Skogestad, Sigurd and Postlethwaite, Ian},
