### 2021-12-05 #97 @gdalle
My second review pass is available here: https://github.com/JuliaReach/LazySets-JuliaCon21/pull/5#issuecomment-986201407

The changes made by the authors significantly improve the clarity of the paper. I still have a handful of very minor formatting issues (see issue link) but I can now recommend **acceptance**
