+++ paper/ref.bib
@@ -263,7 +263,7 @@
 
 @misc{BenetFSS19,
   author = {Benet, Luis and Forets, Marcelo and Sanders, David P. and Schilling, Christian},
-  title = {{TaylorModels}.jl: {T}aylor models in Julia and its application to validated solutions of {ODE}s},
+  title = {{TaylorModels}.jl: {T}aylor models in {J}ulia and its application to validated solutions of {ODE}s},
   booktitle = {{SWIM}},
   year = {2019},
   url = {https://swim2019.ensta-paristech.fr/},
