+++ paper/applications.tex
@@ -163,7 +163,7 @@ H = overapproximate(vTM, Hyperrectangle)
 	\centering
 	\includegraphics[height=3.5cm, keepaspectratio]{img/taylormodel_linear}
 	\includegraphics[height=3.5cm, keepaspectratio]{img/taylormodel_nonlinear}
-	\caption{Left: Exact conversion of a linear Taylor model (red) to a zonotope (green) against an inexact evaluation using interval arithmetic (blue). Right: Approximate conversion of a non-linear Taylor model (red) using a zonotope (green).}
+	\caption{Left: Exact conversion of a linear Taylor model (red) to a zonotope $Z$ (green) against an inexact evaluation $H$ using interval arithmetic (blue, dashed). Right: Approximate conversion of a non-linear Taylor model (red) using a zonotope (green).}
 	\label{fig:taylormodelsconversion}
 \end{figure}
 
