+++ paper/applications.tex
@@ -201,7 +201,7 @@ JuliaReach builds on sound scientific approaches such as \cite{BogomolovFFVPS18}
 
 \smallskip
 
-Reachability techniques are implemented in the JuliaReach package \href{https://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}, which uses LazySets at its core for dealing with sets, including the computation of flowpipes (union of reachable states).
+Reachability techniques are implemented in the JuliaReach package \href{https://github.com/JuliaReach/ReachabilityAnalysis.jl}{ReachabilityAnalysis.jl}, which uses LazySets at its core for dealing with sets, including the computation of reachable states. The union of reachable states for consecutive time intervals is often called a \emph{flowpipe}.
 %
 We include two examples. In Fig.~\ref{fig:flowpipe_heli}, a flowpipe for the vertical velocity of an 8-dimensional helicopter model with a 20-dimensional controller from \cite{skogestad2007multivariable} is shown.
 %
