+++ paper/approx.tex
@@ -89,7 +89,7 @@ It is used via \code{overapproximate(X, $\varepsilon$)}, where $\varepsilon$ is 
 %
 The higher-dimensional extension is not implemented yet.
 %
-On the other hand, higher-dimensional set can be lazily projected using the support function to a lower-dimensional subspace, where the available method applies.
+On the other hand, a higher-dimensional set can be lazily projected using the support function to a lower-dimensional subspace, where the available method applies.
 
 \smallskip
 
