### 2021-12-16 JuliaReach/LazySets-JuliaCon21#4 @schillic
I close this because all items have been addressed by now.
