### 2021-12-17 #97 @matbesancon
@schillic can you confirm that 10.5281/zenodo.5761068 is the DOI of the latest version (which corresponds to the paper)?
