### 2021-12-17 #97 @matbesancon
Also: can you check the missing DOI and add it to the references?
https://github.com/JuliaCon/proceedings-review/issues/97#issuecomment-996584733
