+++ paper/ref.bib
@@ -6,7 +6,8 @@
 	number={6},
 	pages={933--941},
 	year={2008},
-	publisher={Springer}
+	publisher={Springer},
+	doi={10.1134/S0965542508060055}
 }
 
 @article{kamenev1996algorithm,
@@ -72,10 +73,7 @@
 	pages     = {49--75},
 	publisher = {EasyChair},
 	year      = {2020},
-	doi       = {10.29007/zkf6},
-	timestamp = {Sun, 15 Aug 2021 00:16:52 +0200},
-	biburl    = {https://dblp.org/rec/conf/arch/GerettiSABCCCFF20.bib},
-	bibsource = {dblp computer science bibliography, https://dblp.org}
+	doi       = {10.29007/zkf6}
 }
 
 @inproceedings{ARCHCOMP20LINEAR,
@@ -100,10 +98,7 @@
 	pages     = {16--48},
 	publisher = {EasyChair},
 	year      = {2020},
-	doi       = {10.29007/7dt2},
-	timestamp = {Sun, 15 Aug 2021 00:16:52 +0200},
-	biburl    = {https://dblp.org/rec/conf/arch/AlthoffBBFFFKLM20.bib},
-	bibsource = {dblp computer science bibliography, https://dblp.org}
+	doi       = {10.29007/7dt2}
 }
 
 @misc{legat2021polyhedra,
@@ -130,7 +125,7 @@
 	number = {2},
 	pages = {295-320},
 	year = {2017},
-	doi = {10.1137/15M1020575},
+	doi = {10.1137/15M1020575}
 }
 
 @article{MathOptInterface-2021,
@@ -182,12 +177,12 @@
   doi       = {10.1007/978-3-030-89716-1\_6}
 }
 
-@article{schilling2021verification,
+@inproceedings{schilling2021verification,
 	title={Verification of Neural-Network Control Systems by Integrating {T}aylor Models and Zonotopes
 	},
 	author={Schilling, Christian and Forets, Marcelo and Guadalupe, Sebastian},
-	journal={Submitted.},
-	year={2021}
+	booktitle={AAAI},
+	year={2022}
 }
 
 @article{forets2021combining,
