### 2021-12-17 #97 @schillic
@matbesancon
> can you confirm that 10.5281/zenodo.5761068 is the DOI of the latest version (which corresponds to the paper)?

confirmed

> can you check the missing DOI and add it to the references?

Yes, sorry, it's there now!
