### 2021-12-17 #97 @whedon
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
