### 2021-12-17 #97 @mforets
Thank you @matbesancon for the editorial work and @gdalle and @blegat for your time reviewing our submission.

It was our first experience with an open review process. It went so smoothly! We would highly recommend it to our colleagues.
