### 2022-01-27 #97 @matbesancon
@mforets sorry there is a rendering error: https://proceedings.juliacon.org/papers/10.21105/jcon.00097
can you remove the `$^*$` in the title and put the footnote in the text, we'll then update the paper page
