### 2022-01-28 #97 @schillic
@matbesancon we just merged https://github.com/JuliaReach/LazySets-JuliaCon21/pull/7 to remove the star. What should we do now?
